Modularize transfers page
Split Transfers page into per-section components under
src/components/transfers/* (Hero, Types, How, Included, Fleet, Faq, Cta),
lift static content into src/data/transfersPageContent.js, and reuse
usePageMeta. Split src/styles/transfers.css into a manifest of @imports
under src/styles/transfers/* with cascade order preserved.
