+++ src/data/transfersPageContent.js
@@ -1,0 +1,60 @@
+export const TRANSFERS_HERO_IMAGE = '/assets/images/home/mizingani-waterfront.webp';
+export const TRANSFERS_CTA_IMAGE = '/assets/images/home/stone-town-waterfront.webp';
+
+export const TRANSFERS_HOW_IT_WORKS = [
+  {
+    step: '01',
+    title: 'Tell us your details',
+    text: 'Share your flight number, arrival time, hotel name, and group size via WhatsApp or the booking form. We confirm within a few hours.',
+  },
+  {
+    step: '02',
+    title: 'We track your flight',
+    text: "Our drivers monitor your flight in real time. If it's early or delayed, we adjust — no waiting charges, no missed pickups.",
+  },
+  {
+    step: '03',
+    title: 'Seamless arrival',
+    text: "Your driver meets you at arrivals with a name board. Luggage loaded, AC on, you're at your hotel in 20–45 minutes.",
+  },
+];
+
+export const TRANSFERS_INCLUDED = [
+  'Private transfer pricing by route',
+  'Meet & greet with name board',
+  'Flight tracking & delay monitoring',
+  'Air-conditioned vehicle',
+  'Luggage handling',
+  'Bottled water on board for premium bookings',
+  'Tourist-safe licensed drivers',
+  'Hotel drop-off door-to-door',
+  'WhatsApp support before arrival',
+];
+
+export const TRANSFERS_FAQS = [
+  {
+    q: 'How much does an airport transfer cost?',
+    a: 'Private airport transfers start from $25 per vehicle to Stone Town, $45 per vehicle to Paje or Jambiani, and $50 per vehicle to Nungwi or Kendwa. Premium SUV, VIP concierge, group, and minivan rates are priced separately.',
+    open: true,
+  },
+  {
+    q: 'What if my flight is delayed?',
+    a: "We track your flight in real time using the flight number you give us. Your driver will wait at no extra charge for delays up to 2 hours. For longer delays we'll adjust your booking and keep in touch via WhatsApp.",
+  },
+  {
+    q: 'Can you pick me up from my hotel at 3am?',
+    a: 'Yes. We operate 24 hours, 7 days a week. Early morning and late-night pickups are common — just give us the time when you book.',
+  },
+  {
+    q: 'Do you cover the north coast (Nungwi, Kendwa)?',
+    a: 'Yes — we cover the entire island. Stone Town, Nungwi, Kendwa, Matemwe, Kiwengwa, Pongwe, Paje, Bwejuu, Jambiani, and everywhere in between.',
+  },
+  {
+    q: 'Can you arrange transfers for a safari connection?',
+    a: "Absolutely. If you're connecting a Zanzibar stay with a mainland safari, we coordinate the ferry port or Dar airport pickup and can arrange onward road transfer to Arusha or other safari hubs.",
+  },
+  {
+    q: 'How do I pay?',
+    a: 'We accept USD, EUR, GBP, and TZS. Standard private transfers can usually be paid on the day. Premium, VIP, group, and long-distance bookings may require a deposit to secure the vehicle and driver.',
+  },
+];
