+++ src/styles/transfers/hero.css
@@ -1,0 +1,159 @@
+/* ------------ HERO (mirrors .saf-hero / .exc-hero pattern) ------------ */
+.tr-hero {
+  position: relative;
+  min-height: 100svh;
+  display: flex; align-items: center; justify-content: center;
+  text-align: center;
+  padding: clamp(4.75rem, 8vw, 6.5rem) clamp(1.25rem, 5vw, 6rem) clamp(2.25rem, 5vh, 3.5rem);
+  color: #F5F1E8;
+  overflow: hidden;
+  isolation: isolate;
+}
+.tr-hero__bg { position: absolute; inset: 0; z-index: 0; }
+.tr-hero__bg img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  object-position: center 60%;
+  transform: scale(1.05);
+  animation: heroImageReveal 1.4s var(--dp-ease-out) both;
+}
+.tr-hero__bg::after {
+  content: '';
+  position: absolute; inset: 0;
+  background:
+    linear-gradient(120deg, rgba(8,30,46,.85) 0%, rgba(8,30,46,.55) 50%, rgba(8,30,46,.25) 100%),
+    linear-gradient(180deg, transparent 50%, rgba(8,30,46,.7) 100%);
+}
+.tr-hero__inner {
+  position: relative;
+  z-index: 1;
+  width: min(780px, 100%);
+  max-width: 780px;
+  margin: 0 auto;
+  display: grid;
+  justify-items: center;
+  gap: clamp(.75rem, 1.35vh, 1.3rem);
+}
+.tr-hero__eyebrow {
+  display: inline-flex; align-items: center; gap: .5rem;
+  font-family: var(--dp-font-sans);
+  font-size: 11px; letter-spacing: .18em; text-transform: uppercase;
+  background: rgba(255,255,255,.12);
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255,255,255,.18);
+  padding: .5rem 1rem;
+  border-radius: 999px;
+  margin-bottom: 0;
+  animation: fadeInUp .9s var(--dp-ease-out) both;
+}
+.tr-hero__eyebrow::before {
+  content: '';
+  width: 6px; height: 6px;
+  border-radius: 50%;
+  background: var(--dp-accent, #FF6F61);
+  box-shadow: 0 0 10px var(--dp-accent, #FF6F61);
+}
+.tr-hero__title {
+  font-family: var(--dp-font-display);
+  font-size: clamp(2.75rem, 6.5vw, 5.25rem);
+  line-height: 1.02;
+  font-weight: 500;
+  margin: 0;
+  letter-spacing: -.015em;
+  color: #fff;
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .1s;
+}
+.tr-hero__title em {
+  font-style: italic;
+  color: var(--dp-accent);
+  font-weight: 400;
+}
+.tr-hero__lead {
+  font-family: var(--dp-font-display);
+  font-size: clamp(1rem, 1.4vw, 1.2rem);
+  line-height: 1.65;
+  color: rgba(245,241,232,.88);
+  max-width: 640px;
+  margin: 0 auto;
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .3s;
+}
+.tr-hero__cta {
+  display: flex; gap: .75rem; flex-wrap: wrap; justify-content: center;
+  margin-bottom: clamp(.25rem, .7vh, .65rem);
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .5s;
+}
+.tr-hero__stats {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 1.25rem;
+  max-width: 640px;
+  width: 100%;
+  padding-top: clamp(.9rem, 1.7vh, 1.3rem);
+  border-top: 1px solid rgba(255,255,255,.18);
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .7s;
+}
+.tr-hero__stats div { display: flex; flex-direction: column; gap: 4px; }
+.tr-hero__stats strong {
+  font-family: var(--dp-font-display);
+  font-size: clamp(1.5rem, 2.4vw, 2.2rem);
+  font-weight: 500;
+  line-height: 1;
+}
+.tr-hero__stats span {
+  font-family: var(--dp-font-sans);
+  font-size: 11px; letter-spacing: .12em; text-transform: uppercase; opacity: .65;
+}
+
+@media (max-width: 720px) {
+  .tr-hero {
+    min-height: auto;
+    align-items: center;
+    padding: calc(var(--nav-height) + 1.15rem) 1rem 2rem;
+  }
+  .tr-hero__inner { gap: .82rem; }
+  .tr-hero__eyebrow {
+    max-width: min(100%, 23rem);
+    justify-content: center;
+    flex-wrap: wrap;
+    padding: .55rem .78rem;
+    border-radius: 8px;
+    font-size: 9px;
+    line-height: 1.4;
+    letter-spacing: .14em;
+  }
+  .tr-hero__title { font-size: clamp(2.15rem, 11vw, 3.25rem); line-height: 1.02; }
+  .tr-hero__lead { max-width: 24rem; font-size: 1rem; line-height: 1.48; }
+  .tr-hero__cta {
+    width: 100%;
+    flex-direction: column;
+    align-items: center;
+    gap: .65rem;
+    margin-bottom: .25rem;
+  }
+  .tr-hero__cta .btn { width: min(100%, 20rem); min-height: 48px; }
+  .tr-hero__stats {
+    max-width: 22rem;
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    gap: .95rem 1.15rem;
+    padding-top: .95rem;
+  }
+  .tr-hero__stats strong { font-size: clamp(1.35rem, 7vw, 1.7rem); }
+  .tr-hero__stats span { font-size: 9px; letter-spacing: .1em; }
+}
+
+@media (min-width: 721px) and (max-height: 860px) {
+  .tr-hero { padding-top: clamp(4rem, 7vw, 5rem); padding-bottom: 1.75rem; }
+  .tr-hero__title { font-size: clamp(2.35rem, 5.1vw, 4.2rem); }
+  .tr-hero__lead { font-size: clamp(0.95rem, 1.15vw, 1.08rem); line-height: 1.52; }
+  .tr-hero__stats strong { font-size: clamp(1.3rem, 2vw, 1.85rem); }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .tr-hero__bg img { animation: none; transform: scale(1.05); }
+}
+
